requirements.txt
Flask==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
cloudinary==1.44.0
